requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
distlib==0.3.8
Django==5.2
django-cors-headers==4.7.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.16.0
#djangorestframework-jsonapi==6.0.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.15.4
idna==3.10
inflection==0.5.1
mysqlclient==2.2.7
packaging==24.2
pillow==11.0.0
platformdirs==4.2.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0.2
requests==2.32.3
soupsieve==2.6
sqlparse==0.4.3
typing_extensions==4.13.2
tzdata==2022.6
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.26.3